save/restore state system... hope this works :)
[git-p4: depot-paths = "//depot/Golf/": change = 541]

+++ code/game/Classes/RudeRegistryCF.h
@@ -9,8 +9,8 @@ public:
 	RudeRegistryCF(void);
 	~RudeRegistryCF(void);
 
-	long QueryByte(TCHAR *app, TCHAR *name, void *buffer, long *buffersize);
-	long SetByte(TCHAR *app, TCHAR *name, void *buffer, long buffersize);
+	int QueryByte(TCHAR *app, TCHAR *name, void *buffer, int *buffersize);
+	int SetByte(TCHAR *app, TCHAR *name, void *buffer, int buffersize);
 
 private:
